+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Refine pool information content.
+- Limit pool internals public exposure.
 
 ## [2.5.2] - 2023-06-02